add discount on transaction

+++ database/migrations/2022_02_25_092437_create_transaction_details_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->foreignId('product_id')->constrained('products');
             $table->integer('qty');
             $table->double('discount');
-            $table->integer('selling_price');
+            $table->double('selling_price');
             $table->timestamps();
         });
     }
